Resolved circular imports and updated tutorial to accomodate network
