+++ agage_archive/config.yaml
@@ -11,3 +11,9 @@ paths:
     ale_path: ale
     gage_path: gage
     output_path: output
+  agage:
+    agage_md_path: data-nc.zip
+    agage_gcms_path: data-gcms-nc.zip
+    ale_path: ale_gage_sio1993/ale
+    gage_path: ale_gage_sio1993/gage
+    output_path: output
